Extract CloneActionSegment to reduce duplicate logic for MySQL and doris
